intro BaseOptimizeAction
